뷰, 컨트롤러 디렉토리 생성 및 구조 재배치
