+++ src/Controllers/modal.js
@@ -1,0 +1,18 @@
+import {getSearchKeyword} from "./search";
+
+const $searchButton = document.querySelector("#search-button");
+const $modalClose = document.querySelector(".modal-close");
+const $modal = document.querySelector(".modal");
+const $startSearchButton = document.querySelector("#start-search-button");
+
+const onModalShow = () => {
+  $modal.classList.add("open");
+};
+
+const onModalClose = () => {
+  $modal.classList.remove("open");
+};
+
+$searchButton.addEventListener("click", onModalShow);
+$modalClose.addEventListener("click", onModalClose);
+$startSearchButton.addEventListener("click", getSearchKeyword)
