+++ src/Views/index.html
@@ -4,22 +4,22 @@
     <meta charset="UTF-8" />
     <title>👩🏻‍💻 수여닝 유튜브 강의실</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" type="text/css" href="./src/css/index.css" />
-    <link rel="stylesheet" href="./src/css/ui/index.css" />
-    <link rel="stylesheet" href="./src/css/ui/modules/chip.css" />
-    <link rel="stylesheet" href="./src/css/ui/modules/modal.css" />
-    <link rel="stylesheet" href="./src/css/ui/modules/youtubeCard.css" />
-    <link rel="stylesheet" href="./src/css/shared/index.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/background.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/border.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/button.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/flex.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/interactivity.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/layout.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/sizing.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/spacing.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/table.css" />
-    <link rel="stylesheet" href="./src/css/shared/modules/typhography.css" />
+    <link rel="stylesheet" type="text/css" href="../styles/index.css" />
+    <link rel="stylesheet" href="../styles/ui/index.css" />
+    <link rel="stylesheet" href="../styles/ui/modules/chip.css" />
+    <link rel="stylesheet" href="../styles/ui/modules/modal.css" />
+    <link rel="stylesheet" href="../styles/ui/modules/youtubeCard.css" />
+    <link rel="stylesheet" href="../styles/shared/index.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/background.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/border.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/button.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/flex.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/interactivity.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/layout.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/sizing.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/spacing.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/table.css" />
+    <link rel="stylesheet" href="../styles/shared/modules/typhography.css" />
   </head>
   <body>
     <div id="app">
@@ -85,7 +85,7 @@
             <h2 class="text-center">🔎 유튜브 검색</h2>
           </header>
           <form class="d-flex">
-            <input type="text" class="w-100 mr-2 pl-2" placeholder="검색" />
+            <input id="search-input" type="text" class="w-100 mr-2 pl-2" placeholder="검색" />
             <button id="start-search-button" type="button" class="btn bg-cyan-500">검색</button>
           </form>
           <section class="mt-2">
@@ -135,7 +135,9 @@
       </div>
     </div>
 <!--    parcel 환경에서는 type module 로 하면 동작하지 않는다고 함-->
-<!--    <script type="module" src="./src/js/index.js"></script>-->
-    <script src="./src/js/index.js"></script>
+<!--    <script type="module" src="./src/js/modal.js"></script>-->
+<!--    TODO : js 디렉토리 전체를 한번에 임포트하는 방법 찾아보기-->
+    <script src="../Controllers/modal.js"></script>
+    <script src="../Controllers/search.js"></script>
   </body>
 </html>
