+++ src/assets/readme/laptop_with_youtube_logo.png
